update db: change to decimal type for money

+++ DTO/DiscountDTO/Discount.CreateDTO.cs
@@ -8,6 +8,6 @@ namespace uni_cap_pro_be.DTO.DiscountDTO
 		[Required]
 		public required short Level { get; set; }
 		[Required]
-		public required double Amount { get; set; }
+		public required float Amount { get; set; }
 	}
 }
