+++ Models/Order.Model.cs
@@ -13,7 +13,7 @@ namespace uni_cap_pro_be.Models
 		public DateTime Modified_At { get; set; }
 
 		[Required]
-		public required double Total_Price { get; set; }
+		public required decimal Total_Price { get; set; }
 
 		[Required]
 		public required int Total_Quantity { get; set; }
